+++ internal/search/openserp/embedded_darwin_amd64.go
@@ -1,0 +1,8 @@
+//go:build openserp_embedded && darwin && amd64
+
+package openserp
+
+import _ "embed"
+
+//go:embed assets/openserp-darwin-amd64
+var embeddedBinary []byte