+++ internal/search/openserp/embedded_default.go
@@ -1,0 +1,5 @@
+//go:build !openserp_embedded
+
+package openserp
+
+var embeddedBinary []byte