+++ internal/search/openserp/embedded_linux_amd64.go
@@ -1,0 +1,8 @@
+//go:build openserp_embedded && linux && amd64
+
+package openserp
+
+import _ "embed"
+
+//go:embed assets/openserp-linux-amd64
+var embeddedBinary []byte