+++ internal/search/openserp/embedded_windows_386.go
@@ -1,0 +1,8 @@
+//go:build openserp_embedded && windows && 386
+
+package openserp
+
+import _ "embed"
+
+//go:embed assets/openserp-windows-386
+var embeddedBinary []byte