+++ internal/search/openserp/embedded_windows_amd64.go
@@ -1,0 +1,8 @@
+//go:build openserp_embedded && windows && amd64
+
+package openserp
+
+import _ "embed"
+
+//go:embed assets/openserp-windows-amd64
+var embeddedBinary []byte